new bits

+++ src/reencode.rs
@@ -1,3 +1,4 @@
+use crate::scan::file::ScannedFile;
 use postgres::Connection;
 use std::fs;
 use std::path::Path;
@@ -12,8 +13,6 @@ impl crate::module::Module for Reencode {
     fn module_iteration(&self, connection: &Connection) -> () {
         info!("Searching for targets to reencode");
         let mut done = false;
-        let mut offset: i32 = 0;
-        let limit: i32 = 100;
         let target_extension = self.config_string(&connection, "target_extension");
         let target_codec = self.config_string(&connection, "target_codec");
         while !done {
@@ -22,21 +21,37 @@ impl crate::module::Module for Reencode {
                 "Selecting paths where extension+codec do not match {}+{}",
                 &target_extension, &target_codec
             );
-            let rows = connection.query("SELECT path FROM paths INNER JOIN video_extensions USING(extension) WHERE extension != $1 or codec != $2 ORDER BY path DESC OFFSET $3::int4 LIMIT $4::int4", &[&target_extension, &target_codec, &offset, &limit]).unwrap();
+            let rows = connection
+                .query(
+                    "\
+                UPDATE paths SET in_progress = true WHERE id = ( \
+                    SELECT id FROM paths \
+                    INNER JOIN video_extensions USING(extension) \
+                    WHERE (extension != $1 or codec != $2) AND NOT in_progress \
+                    LIMIT 1 \
+                )\
+                RETURNING id, path, bytes
+                ",
+                    &[&target_extension, &target_codec],
+                )
+                .unwrap();
             debug!("Got {:?}", rows);
-            offset += limit;
             for row in rows.iter() {
+                // always just one unless its zero
                 done = false;
-                let source_path: String = row.get(0);
-                let target_path = Path::new(&source_path).with_extension(&target_extension);
+                let id: i64 = row.get(0);
+                let source_path_s: String = row.get(1);
+                let source_path = Path::new(&source_path_s);
+                let original_bytes: i64 = row.get(2);
+                let target_path = source_path.with_extension(&target_extension);
                 let temp_path = Path::new("/tmp/converting.x").with_extension(&target_extension);
-                info!("Converting {}", &source_path);
+                info!("Converting {:?}", &source_path);
                 let captured = Exec::cmd("ffmpeg")
                     .arg("-y")
                     .arg("-loglevel")
                     .arg("warning")
                     .arg("-i")
-                    .arg(&source_path)
+                    .arg(&source_path_s) // danger!
                     .arg("-c:v")
                     .arg(&target_codec)
                     .arg("-c:a")
@@ -51,8 +66,18 @@ impl crate::module::Module for Reencode {
                 if captured.success() {
                     info!("cp {:?} {:?}", &temp_path, &target_path);
                     fs::copy(&temp_path, &target_path).unwrap();
-                    info!("rm {}", &source_path);
-                    fs::remove_file(&source_path).unwrap();
+                    let new_file = ScannedFile::new(&target_path, &connection).unwrap();
+                    info!(
+                        "Bytes {:?} -> {:?} = {:?}",
+                        original_bytes,
+                        new_file.bytes,
+                        new_file.bytes - original_bytes
+                    );
+                    new_file.store(&connection);
+                    if source_path != target_path {
+                        info!("rm {:?}", &source_path);
+                        fs::remove_file(&source_path).unwrap();
+                    }
                 } else {
                     warn!("ffmpeg failed: {}", &captured.stderr_str());
                 }
